// src/types.ts
export type MenuItem =
  | "AboutPhone"
  | "SystemAppsUpdater"
  | "SecurityStatus"
  | "SimCardsMobileNetworks"
  | "WiFi"
  | "Bluetooth"
  | "ConnectionSharing"
  | "LockScreen"
  | "Display"
  | "SoundVibration"
  | "Notification"
  | "HomeScreen"
  | "Wallpaper"
  | "Themes"
  | "PasswordsSecurity"
  | "PrivacyProtection"
  | "BatteryPerformance"
  | "Apps"
  | "AdditionalSettings"
  | "DigitalWellbeingParentalControls"
  | "SpecialFeatures"
  | "Google"
  | "AccountSync"
  | "ServicesFeedback"